Move resign_mode state management into CancelButton.
