+++ src/views/Game/Game.tsx
@@ -131,8 +131,6 @@ export function Game(): JSX.Element {
     const [title, set_title] = React.useState<string>();
 
     const [mode, set_mode] = React.useState<GobanModes>("play");
-    const [resign_mode, set_resign_mode] = React.useState<"cancel" | "resign">();
-    const [resign_text, set_resign_text] = React.useState<string>();
     const [score_estimate_winner, set_score_estimate_winner] = React.useState<string>();
     const [score_estimate_amount, set_score_estimate_amount] = React.useState<number>();
     const [show_title, set_show_title] = React.useState<boolean>();
@@ -755,42 +753,6 @@ export function Game(): JSX.Element {
         return ret;
     };
 
-    const cancelOrResign = () => {
-        let dropping_from_casual_rengo = false;
-
-        if (goban.current.engine.rengo && goban.current.engine.rengo_casual_mode) {
-            const team = goban.current.engine.rengo_teams.black.find(
-                (p) => p.id === data.get("user").id,
-            )
-                ? "black"
-                : "white";
-            dropping_from_casual_rengo = goban.current.engine.rengo_teams[team].length > 1;
-        }
-
-        if (resign_mode === "cancel") {
-            swal({
-                text: _("Are you sure you wish to cancel this game?"),
-                confirmButtonText: _("Yes"),
-                cancelButtonText: _("No"),
-                showCancelButton: true,
-                focusCancel: true,
-            })
-                .then(() => goban.current.cancelGame())
-                .catch(() => 0);
-        } else {
-            swal({
-                text: dropping_from_casual_rengo
-                    ? _("Are you sure you want to abandon your team?")
-                    : _("Are you sure you wish to resign this game?"),
-                confirmButtonText: _("Yes"),
-                cancelButtonText: _("No"),
-                showCancelButton: true,
-                focusCancel: true,
-            })
-                .then(() => goban.current.resign())
-                .catch(() => 0);
-        }
-    };
     const goban_setModeDeferredPlay = () => {
         goban.current.setModeDeferred("play");
     };
@@ -847,9 +809,6 @@ export function Game(): JSX.Element {
             goban={goban.current}
             show_cancel={show_cancel}
             player_to_move={player_to_move}
-            onCancel={cancelOrResign}
-            resign_text={resign_text}
-            view_mode={view_mode}
             user_is_player={user_is_player}
             review_list={review_list}
             stashed_conditional_moves={stashed_conditional_moves.current}
@@ -1208,16 +1167,6 @@ export function Game(): JSX.Element {
 
         /* Ensure our state is kept up to date */
 
-        const sync_resign_text = () => {
-            if (goban.current.engine.gameCanBeCanceled()) {
-                set_resign_text(_("Cancel game"));
-                set_resign_mode("cancel");
-            } else {
-                set_resign_text(_("Resign"));
-                set_resign_mode("resign");
-            }
-        };
-
         const sync_show_title = () =>
             set_show_title(
                 !goban.current.submit_move ||
@@ -1258,7 +1207,6 @@ export function Game(): JSX.Element {
             set_score_estimate_winner(undefined);
             set_undo_requested(engine.undo_requested);
 
-            sync_resign_text();
             sync_show_title();
             sync_move_info();
             sync_stone_removal();
@@ -1304,7 +1252,6 @@ export function Game(): JSX.Element {
             set_score_estimate_amount(est?.amount);
         });
         goban.current.on("undo_requested", set_undo_requested);
-        goban.current.on("cur_move", sync_resign_text);
         goban.current.on("cur_move", sync_show_title);
         goban.current.on("submit_move", sync_show_title);
         goban.current.on("cur_move", sync_move_info);
@@ -1643,11 +1590,7 @@ export function Game(): JSX.Element {
                         !zen_mode &&
                         user_is_player &&
                         phase !== "finished") ||
-                        null) && (
-                        <CancelButton view_mode={view_mode} onClick={cancelOrResign}>
-                            {resign_text}
-                        </CancelButton>
-                    )}
+                        null) && <CancelButton goban={goban.current} className="bold reject" />}
 
                     {((view_mode === "portrait" && !zen_mode) || null) && (
                         <GameDock
